WT-3348 Lint, Windows warnings. (#3449)
* Simplify Windows include files, clean up Windows warnings.

* Remove incorrect path from copyright skip list.

* Clean up a set of places where an operation could potentially truncate a value by performing the operation at the larger size and then casting the result.

+++ src/block/block_ext.c
@@ -1272,7 +1272,7 @@ __wt_block_extlist_write(WT_SESSION_IMPL *session,
 	 * entries: the initial WT_BLOCK_EXTLIST_MAGIC/0 pair and the list-
 	 * terminating WT_BLOCK_INVALID_OFFSET/0 pair.
 	 */
-	size = (entries + 2) * 2 * WT_INTPACK64_MAXSIZE;
+	size = ((size_t)entries + 2) * 2 * WT_INTPACK64_MAXSIZE;
 	WT_RET(__wt_block_write_size(session, block, &size));
 	WT_RET(__wt_scr_alloc(session, size, &tmp));
 	dsk = tmp->mem;
